+++ packages/design-system/src/components/button-icon/button-icon.tsx
@@ -1,0 +1,17 @@
+import IconButton from '@mui/material/IconButton';
+
+interface Props {
+  onClick?: () => void;
+  ariaLabel: string;
+  icon: React.ReactNode;
+}
+
+const ButtonIcon = ({ onClick, ariaLabel, icon }: Props) => {
+  return (
+    <IconButton onClick={onClick} aria-label={ariaLabel} disableRipple>
+      {icon}
+    </IconButton>
+  );
+};
+
+export default ButtonIcon;
